rename content directory to homepage
