+++ frontend/src/components/homepage/BuildingContent.tsx
