+++ frontend/src/components/homepage/InvolvedContent.tsx
