+++ frontend/stories/components/GoalContent.stories.tsx
@@ -1,6 +1,6 @@
-import { Meta } from "@storybook/react";
+import GoalContent from "src/components/homepage/ExperimentalContent";
 
-import GoalContent from "src/components/content/ExperimentalContent";
+import { Meta } from "@storybook/react";
 
 const meta: Meta<typeof GoalContent> = {
   title: "Components/Content/Goal Content",
