+++ frontend/tests/components/GoalContent.test.tsx
@@ -1,6 +1,8 @@
-import { render, screen } from "tests/react-utils";
-
-import GoalContent from "src/components/content/ExperimentalContent";
+import GoalContent from "src/components/homepage/ExperimentalContent";
+import {
+  render,
+  screen,
+} from "tests/react-utils";
 
 describe("Goal Content", () => {
   it("Renders without errors", () => {
